requirements.txt
python-telegram-bot==20.3
openai==0.27.0
python-dotenv==1.0.0
SQLAlchemy==1.4.46
psycopg2-binary==2.9.6
librosa==0.9.2
SpeechRecognition==3.8.1
soundfile==0.12.1
matplotlib==3.10.0
psycopg2==2.9.10
setuptools==74.1.2
numpy==1.23.5
pandas==1.5.3
openpyxl==3.1.2
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.159.0
plotly==5.15.0
kaleido==0.2.1
fonttools==4.55.6
APScheduler==3.11.0
